+++ idle-application/src/main/kotlin/com/swm/idle/application/user/carer/facade/CarerAuthFacadeService.kt
@@ -15,7 +15,6 @@ import com.swm.idle.domain.user.common.vo.BirthYear
 import com.swm.idle.domain.user.common.vo.PhoneNumber
 import com.swm.idle.infrastructure.client.geocode.service.GeoCodeService
 import com.swm.idle.support.security.exception.SecurityException
-import com.swm.idle.support.transfer.auth.center.RefreshLoginTokenResponse
 import com.swm.idle.support.transfer.auth.common.LoginResponse
 import org.springframework.stereotype.Service
 
@@ -100,16 +99,6 @@ class CarerAuthFacadeService(
             role = UserType.CARER,
             reason = reason,
         )
-    }
-
-    fun refreshLoginToken(refreshToken: String): RefreshLoginTokenResponse {
-        return refreshTokenService.create(refreshToken)
-            .let {
-                RefreshLoginTokenResponse(
-                    accessToken = it.accessToken,
-                    refreshToken = it.refreshToken,
-                )
-            }
     }
 
 }
