+++ idle-application/src/main/kotlin/com/swm/idle/application/user/common/service/facade/AuthFacadeService.kt
@@ -1,10 +1,12 @@
 package com.swm.idle.application.user.common.service.facade
 
+import com.swm.idle.application.user.common.service.domain.RefreshTokenService
 import com.swm.idle.application.user.common.service.domain.UserPhoneVerificationService
 import com.swm.idle.application.user.vo.UserPhoneVerificationNumber
 import com.swm.idle.domain.user.common.exception.UserException
 import com.swm.idle.domain.user.common.vo.PhoneNumber
 import com.swm.idle.infrastructure.sms.auth.service.SmsService
+import com.swm.idle.support.transfer.auth.center.RefreshLoginTokenResponse
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -15,6 +17,7 @@ import org.springframework.transaction.annotation.Transactional
 class AuthFacadeService(
     private val smsService: SmsService,
     private val userPhoneVerificationService: UserPhoneVerificationService,
+    private val refreshTokenService: RefreshTokenService,
 ) {
 
     @Transactional
@@ -45,6 +48,17 @@ class AuthFacadeService(
                 throw UserException.InvalidVerificationNumber()
             }
         } ?: throw UserException.VerificationNumberNotFound()
+    }
+
+    @Transactional
+    fun refreshLoginToken(refreshToken: String): RefreshLoginTokenResponse {
+        return refreshTokenService.create(refreshToken)
+            .let {
+                RefreshLoginTokenResponse(
+                    accessToken = it.accessToken,
+                    refreshToken = it.refreshToken,
+                )
+            }
     }
 
 }
