+++ idle-presentation/src/main/kotlin/com/swm/idle/presentation/auth/carer/controller/CarerAuthController.kt
@@ -8,8 +8,6 @@ import com.swm.idle.presentation.auth.carer.api.CarerAuthApi
 import com.swm.idle.support.transfer.auth.carer.CarerJoinRequest
 import com.swm.idle.support.transfer.auth.carer.CarerLoginRequest
 import com.swm.idle.support.transfer.auth.carer.CarerWithdrawRequest
-import com.swm.idle.support.transfer.auth.center.RefreshLoginTokenResponse
-import com.swm.idle.support.transfer.auth.center.RefreshTokenRequest
 import com.swm.idle.support.transfer.auth.common.LoginResponse
 import org.springframework.web.bind.annotation.RestController
 
@@ -42,10 +40,6 @@ class CarerAuthController(
 
     override fun withdraw(request: CarerWithdrawRequest) {
         carerAuthFacadeService.withdraw(request.reason)
-    }
-
-    override fun refreshLoginToken(request: RefreshTokenRequest): RefreshLoginTokenResponse {
-        return carerAuthFacadeService.refreshLoginToken(request.refreshToken)
     }
 
 }
