+++ idle-presentation/src/main/kotlin/com/swm/idle/presentation/auth/common/api/AuthApi.kt
@@ -1,6 +1,8 @@
 package com.swm.idle.presentation.auth.common.api
 
 import com.swm.idle.presentation.common.exception.ErrorResponse
+import com.swm.idle.support.transfer.auth.center.RefreshLoginTokenResponse
+import com.swm.idle.support.transfer.auth.center.RefreshTokenRequest
 import com.swm.idle.support.transfer.auth.common.ConfirmSmsVerificationRequest
 import com.swm.idle.support.transfer.auth.common.SendSmsVerificationRequest
 import io.swagger.v3.oas.annotations.Operation
@@ -72,5 +74,12 @@ interface AuthApi {
     fun confirmVerificationMessage(
         @RequestBody request: ConfirmSmsVerificationRequest,
     )
+
+    @Operation(summary = "Refresh Login Token API")
+    @PostMapping("/refresh")
+    @ResponseStatus(HttpStatus.OK)
+    fun refreshLoginToken(
+        @RequestBody request: RefreshTokenRequest,
+    ): RefreshLoginTokenResponse
 
 }
